pyproject.toml
[tool.poetry]
name = "scamshield"
version = "0.1.0"
description = ""
authors = ["DACHXY <Danny10132024@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.14.3"
pydantic = "^2.6.4"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
uuid = "^1.30"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
bs4 = "^0.0.2"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
websockets = "^12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
